# 07_API_CONTRACT.yaml
openapi: 3.1.0
info:
  title: Workshops Platform API
  version: 1.0.0
  description: >-
    العقد المرجعي لمنصة ورش. المصادقة نفسها تحت /api/auth/* تديرها Better Auth؛
    بقية المسارات تحت /api/v1 وتستخدم جلسة Cookie آمنة وCSRF عند تغيير الحالة.
servers:
  - url: https://workshops.gumaan.app/api/v1
    description: Production
  - url: http://localhost:5173/api/v1
    description: Local
...
security:
  - sessionCookie: []
paths:
...
  /workspaces:
    get:
      tags: [Workspaces]
      operationId: listMyWorkspaces
      responses:
        "200": {$ref: "#/components/responses/WorkspaceList"}
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      x-permission: platform_admin or account workspace creation policy
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WorkspaceInput"}
      responses:
        "201": {$ref: "#/components/responses/WorkspaceItem"}
        "422": {$ref: "#/components/responses/ValidationError"}
  /workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspaces]
      operationId: getWorkspace
      x-permission: workspace.read
      responses:
        "200": {$ref: "#/components/responses/WorkspaceItem"}
        "404": {$ref: "#/components/responses/NotFound"}
    patch:
      tags: [Workspaces]
      operationId: updateWorkspace
      x-permission: workspace.update
      parameters:
        - $ref: "#/components/parameters/IfMatch"
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/WorkspacePatch"}
      responses:
        "200": {$ref: "#/components/responses/WorkspaceItem"}
        "409": {$ref: "#/components/responses/Conflict"}
        "422": {$ref: "#/components/responses/ValidationError"}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: __Host-workshops_session
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: {$ref: "#/components/schemas/Id"}
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: {type: string}
      description: Entity version or ETag used for optimistic concurrency.
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: {type: string, minLength: 16, maxLength: 128}
  schemas:
    Id:
      type: string
      minLength: 8
      maxLength: 64
      pattern: "^[A-Za-z0-9_-]+$"
...
    Error:
      type: object
      required: [code, message, requestId]
      additionalProperties: false
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: object
          additionalProperties:
            type: array
            items: {type: string}
        requestId: {type: string}
        details: {type: object, additionalProperties: true}
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error: {$ref: "#/components/schemas/Error"}
    PageMeta:
      type: object
      required: [hasMore]
      properties:
        nextCursor: {type: [string, "null"]}
        hasMore: {type: boolean}
        total: {type: integer, minimum: 0}
...
    Workspace:
      type: object
      required: [id, slug, name, defaultLocale, timezone, status, version]
      properties:
        id: {$ref: "#/components/schemas/Id"}
        slug: {type: string}
        name: {type: string}
        description: {type: [string, "null"]}
        primaryColor: {type: string}
        accentColor: {type: string}
        defaultLocale: {type: string, enum: [ar, en]}
        timezone: {type: string}
        status: {type: string, enum: [active, suspended, pending_deletion]}
        myRole: {type: string}
        version: {type: integer, minimum: 1}
    WorkspaceInput:
      type: object
      required: [slug, name, timezone]
      additionalProperties: false
      properties:
        slug: {type: string, pattern: "^[a-z0-9-]{3,50}$"}
        name: {type: string, minLength: 2, maxLength: 160}
        description: {type: string, maxLength: 2000}
        timezone: {type: string, default: Asia/Aden}
        defaultLocale: {type: string, enum: [ar, en], default: ar}
    WorkspacePatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        slug: {type: string, pattern: "^[a-z0-9-]{3,50}$"}
        name: {type: string, minLength: 2, maxLength: 160}
        description: {type: string, maxLength: 2000}
        timezone: {type: string}
        defaultLocale: {type: string, enum: [ar, en]}
        primaryColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        accentColor: {type: string, pattern: "^#[0-9A-Fa-f]{6}$"}
        retentionDays: {type: integer, minimum: 30}
...
    WorkspaceEnvelope:
      type: object
      properties:
        data: {$ref: "#/components/schemas/Workspace"}
    PaginatedWorkspaces:
      type: object
      properties:
        data: {type: array, items: {$ref: "#/components/schemas/Workspace"}}
        meta: {$ref: "#/components/schemas/PageMeta"}
...
  responses:
...
    NotFound:
      description: Resource not found in the authorized scope.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    Conflict:
      description: State transition, duplicate, or optimistic concurrency conflict.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
    ValidationError:
      description: Input failed schema or business validation.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/ErrorEnvelope"}
...
    WorkspaceItem:
      description: Workspace.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/WorkspaceEnvelope"}
    WorkspaceList:
      description: Workspaces.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/PaginatedWorkspaces"}
...